move app tagging to AppInstallAspect
