+++ src/main/java/org/commcare/formplayer/aspects/AppInstallAspect.java
@@ -1,14 +1,21 @@
 package org.commcare.formplayer.aspects;
 
+import io.sentry.Sentry;
+import lombok.extern.java.Log;
+import org.apache.http.client.utils.URIBuilder;
 import org.commcare.formplayer.beans.InstallRequestBean;
 import org.aspectj.lang.JoinPoint;
 import org.aspectj.lang.annotation.Aspect;
 import org.aspectj.lang.annotation.Before;
+import org.commcare.formplayer.util.Constants;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.annotation.Order;
 import org.commcare.formplayer.services.FormplayerStorageFactory;
 import org.commcare.formplayer.util.FormplayerSentry;
 
+import javax.validation.constraints.NotNull;
+import java.net.URISyntaxException;
 import java.util.Arrays;
 
 /**
@@ -16,6 +23,7 @@ import java.util.Arrays;
  */
 @Aspect
 @Order(6)
+@Log
 public class AppInstallAspect {
 
     @Autowired
@@ -23,6 +31,9 @@ public class AppInstallAspect {
 
     @Autowired
     private FormplayerSentry raven;
+
+    @Value("${commcarehq.host}")
+    private String host;
 
     @Before(value = "@annotation(org.commcare.formplayer.annotations.AppInstall)")
     public void configureStorageFactory(JoinPoint joinPoint) throws Throwable {
@@ -42,6 +53,31 @@ public class AppInstallAspect {
                 .setCategory("app_info")
                 .record();
 
-        raven.configureAppUrls(requestBean.getDomain(), requestBean.getAppId());
+        configureAppUrls(requestBean.getDomain(), requestBean.getAppId());
+    }
+
+    public void configureAppUrls(String domain, String appId) {
+        Sentry.configureScope(scope -> {
+            scope.setExtra(Constants.APP_DOWNLOAD_URL_EXTRA, getAppDownloadURL(domain, appId));
+            scope.setExtra(Constants.APP_URL_EXTRA, getAppURL(domain, appId));
+        });
+    }
+
+    private String getAppURL(String domain, String appId) {
+        return host + "/a/" + domain + "/apps/view/" + appId + "/";
+    }
+
+    private String getAppDownloadURL(String domain, String appId) {
+        String baseURL = host + "/a/" + domain + "/apps/api/download_ccz/";
+        URIBuilder builder;
+        try {
+            builder = new URIBuilder(baseURL);
+        } catch (URISyntaxException e) {
+            log.info("Unable to build app download URL");
+            return "unknown";
+        }
+        builder.addParameter("app_id", appId);
+        builder.addParameter("latest", Constants.CCZ_LATEST_SAVED);
+        return builder.toString();
     }
 }
